requirements.txt
django-bootstrap3
django-bootstrap4
